refactor: posistion of the floating button design code
