--- packages/app-installation-cta/src/floating-button-cta-v2/index.ts
@@ -1,3 +1,0 @@
-import FloatingButtonCtaV2 from './floating-button-cta-v2'
-
-export default FloatingButtonCtaV2
