+++ packages/app-installation-cta/src/index.tsx
@@ -2,7 +2,6 @@ import ImageBanner from './image-banner'
 import TextBanner from './text-banner'
 import AppInstallationCTA from './app-installation-cta'
 import BannerCTA from './banner-cta'
-import FloatingButtonCtaV2 from './floating-button-cta-v2'
 import FloatingButtonCTA from './floating-button-cta'
 import ChatbotCTA from './chatbot-cta'
 import { fetchInventoryItems } from './service'
@@ -15,7 +14,6 @@ export {
   AppInstallationCTA,
   BannerCTA,
   FloatingButtonCTA,
-  FloatingButtonCtaV2,
   ChatbotCTA,
   ArticleCardCTA,
   fetchInventoryItems,
